Move page tests into tests/pages/.
