+++ frontend/lib/tests/pages/not-found.test.tsx
@@ -1,4 +1,4 @@
-import { NotFound } from '../pages/not-found';
+import { NotFound } from '../../pages/not-found';
 
 test('NotFound sets status on static renders', () => {
   const ctx: any = { staticContext: {} };
